Split shared.py into barrier.py, memory.py, and prefix.py.
